chore(simulator): move utils to separate folder
