+++ lib/simulator/behaviors/ActivityBehavior.js
@@ -1,7 +1,7 @@
 import {
   isMessageFlow,
   isSequenceFlow
-} from './ModelUtil';
+} from '../util/ModelUtil';
 
 
 export default function ActivityBehavior(simulator, scopeBehavior) {
