+++ lib/simulator/behaviors/BoundaryEventBehavior.js
@@ -1,6 +1,6 @@
 import {
   getBusinessObject
-} from './ModelUtil';
+} from '../util/ModelUtil';
 
 
 export default function BoundaryEventBehavior(
