+++ lib/simulator/behaviors/EndEventBehavior.js
@@ -1,6 +1,6 @@
 import {
   isMessageFlow
-} from './ModelUtil';
+} from '../util/ModelUtil';
 
 
 export default function EndEventBehavior(
