+++ lib/simulator/behaviors/EventBehaviors.js
@@ -7,7 +7,7 @@ import {
 
 import {
   isEventSubProcess
-} from './ModelUtil';
+} from '../util/ModelUtil';
 
 
 export default function EventBehaviors(
