+++ lib/simulator/behaviors/ExclusiveGatewayBehavior.js
@@ -1,6 +1,6 @@
 import {
   filterSequenceFlows
-} from './ModelUtil';
+} from '../util/ModelUtil';
 
 
 export default function ExclusiveGatewayBehavior(simulator) {
