+++ lib/simulator/behaviors/IntermediateCatchEventBehavior.js
@@ -1,6 +1,6 @@
 import {
   isLink
-} from './ModelUtil';
+} from '../util/ModelUtil';
 
 
 export default function IntermediateCatchEventBehavior(
