+++ lib/simulator/behaviors/ProcessBehavior.js
@@ -1,6 +1,6 @@
 import {
   is
-} from './ModelUtil';
+} from '../util/ModelUtil';
 
 
 export default function ProcessBehavior(
