+++ lib/simulator/behaviors/SubProcessBehavior.js
@@ -2,7 +2,7 @@ import {
   is,
   isEventSubProcess,
   isInterrupting
-} from './ModelUtil';
+} from '../util/ModelUtil';
 
 
 export default function SubProcessBehavior(
