+++ lib/simulator/util/ModelUtil.js
